+++ src/internal/types/parser.ts
@@ -1,39 +1,14 @@
-/** Option Types */
-export interface FileOptions {
-	entry: string;
-	minimal?: boolean;
-	exclude?: Array<string>;
-}
-export interface DirOptions extends FileOptions {
-	entry: string;
-	recurse?: boolean;
-	extensions?: Array<string>;
-}
-export interface ParserTypes<Input, Output> {
-	Symbol(): Input;
-	Symbol(): Output;
-}
+import type { MarquaData, MarquaTable } from './data';
+import type { FileOptions } from './option';
 
-/** Data Types */
-export interface MarquaData {
-	type: string;
-	title: string;
-	body: string | Array<this>;
-}
-export interface MarquaTable {
-	id: string;
-	title: string;
-	sections?: Array<this>;
-}
-
-/** Parser Types */
-export interface FrontMatter {
+export type FrontMatter = {
 	toc: Array<MarquaTable>;
 	read_time: number;
 	content?: string | Array<MarquaData>;
 	date: Record<'created' | 'modified', Date> &
 		Partial<Record<'published' | 'updated', string | Date>>;
-}
+};
+
 export type HydrateFn<Options extends FileOptions, Input, Output = Input> = (chunk: {
 	frontMatter: keyof Input extends never
 		? Options['minimal'] extends true